First step to deal with high resolution displays (like Mac Retina): add new screenToViewport/viewportToScreen GlMainWidget member functions + add new viewportTo3DWorld/worldTo2DViewport camera member functions (will deprecated the old screenTo3DWorld/worldTo2DSCreen functions) + updates tulip-ogl and tulip-gui classes according to this
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@11090 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip-gui/include/tulip/MouseEdgeBendEditor.h
@@ -130,9 +130,9 @@ private:
   bool haveSelection(GlMainWidget *);
   void computeSrcTgtEntities(GlMainWidget *);
   bool computeBendsCircles(GlMainWidget*);
-  void mMouseTranslate(double, double, GlMainWidget*);
+  void mMouseTranslate(int, int, GlMainWidget*);
   void mMouseDelete();
-  void mMouseCreate(double, double, GlMainWidget*);
+  void mMouseCreate(int, int, GlMainWidget*);
 };
 
 }
